Moved FrameInfo into wasmer types
