+++ lib/compiler/src/engine/trap/mod.rs
@@ -2,6 +2,5 @@ mod error;
 mod frame_info;
 pub use error::RuntimeError;
 pub use frame_info::{
-    register as register_frame_info, FrameInfo, FunctionExtent, GlobalFrameInfoRegistration,
-    FRAME_INFO,
+    register as register_frame_info, FunctionExtent, GlobalFrameInfoRegistration, FRAME_INFO,
 };
