+++ lib/types/src/compilation/mod.rs
@@ -5,8 +5,6 @@ pub mod function;
 pub mod module;
 pub mod relocation;
 pub mod section;
-pub mod sourceloc;
 pub mod symbols;
 pub mod target;
-pub mod trap;
 pub mod unwind;
