+++ lib/types/src/stack/mod.rs
@@ -1,0 +1,9 @@
+//! Types for the stack tracing / frames.
+
+mod frame;
+mod sourceloc;
+mod trap;
+
+pub use frame::FrameInfo;
+pub use sourceloc::SourceLoc;
+pub use trap::TrapInformation;
