+++ lib/types/src/stack/trap.rs
